Move functionality into account.py
The `loans` logic is in `account.py`. Move tho `loan-history` logic
there as well.
