+++ openlibrary/plugins/upstream/account.py
@@ -2,7 +2,7 @@ from datetime import datetime
 import json
 import logging
 import re
-from typing import Any
+from typing import Any, Final
 from collections.abc import Callable
 from collections.abc import Iterable, Mapping
 
@@ -25,6 +25,10 @@ from openlibrary.core import stats
 from openlibrary.core import helpers as h, lending
 from openlibrary.core.booknotes import Booknotes
 from openlibrary.core.bookshelves import Bookshelves
+from openlibrary.core.lending import (
+    get_availability_of_ocaids,
+    s3_loan_api,
+)
 from openlibrary.core.observations import Observations
 from openlibrary.core.ratings import Ratings
 from openlibrary.plugins.recaptcha import recaptcha
@@ -43,7 +47,9 @@ from openlibrary.utils.dateutil import elapsed_time
 
 logger = logging.getLogger("openlibrary.account")
 
+CONFIG_IA_DOMAIN: Final = config.get('ia_base_url', 'https://archive.org')
 USERNAME_RETRIES = 3
+RESULTS_PER_PAGE: Final = 25
 
 # XXX: These need to be cleaned up
 send_verification_email = accounts.send_verification_email
@@ -1077,6 +1083,41 @@ class account_loans_json(delegate.page):
         return delegate.RawText(json.dumps({"loans": loans}))
 
 
+class account_loan_history(delegate.page):
+    path = "/people/([^/]+)/loan-history"
+
+    @require_login
+    def GET(self, username):
+        i = web.input(page=1)
+        page = int(i.page)
+        mb = MyBooksTemplate(username, key='loan_history')
+        if mb.is_my_page:
+            loan_history_data = get_loan_history_data(page=page, mb=mb)
+            template = render['account/loan_history'](
+                docs=loan_history_data['docs'],
+                current_page=page,
+                show_next=loan_history_data['show_next'],
+                ia_base_url=CONFIG_IA_DOMAIN,
+            )
+            return mb.render(header_title=_("Loan History"), template=template)
+        raise web.seeother(mb.user.key)
+
+
+class account_loan_history_json(delegate.page):
+    encoding = "json"
+    path = "/people/([^/]+)/loan-history"
+
+    @require_login
+    def GET(self, username):
+        i = web.input(page=1)
+        page = int(i.page)
+        mb = MyBooksTemplate(username, key='loan_history')
+        if mb.is_my_page:
+            loan_history_data = get_loan_history_data(page=page, mb=mb)
+            web.header('Content-Type', 'application/json')
+            return delegate.RawText(json.dumps({"loans_history": loan_history_data}))
+
+
 class account_waitlist(delegate.page):
     path = "/account/waitlist"
 
@@ -1140,3 +1181,83 @@ def process_goodreads_csv(i):
         else:
             books_wo_isbns[_book['Book Id']] = _book
     return books, books_wo_isbns
+
+
+def get_loan_history_data(page: int, mb: "MyBooksTemplate") -> dict[str, str | int]:
+    """
+    Retrieve IA loan history data for page `page` of the patron's history.
+
+    This will use a patron's S3 keys to query the IA loan history API,
+    get the IA IDs, get the OLIDs if available, and and then convert this
+    into editions and IA-only items for display in the loan history.
+
+    This returns both editions and IA-only items because the loan history API
+    includes items that are not in Open Library, and displaying only IA
+    items creates pagination and navigation issues. For further discussion,
+    see https://github.com/internetarchive/openlibrary/pull/8375.
+    """
+    if not (account := OpenLibraryAccount.get(username=mb.username)):
+        raise render.notfound(
+            "Account for not found for %s" % mb.username, create=False
+        )
+    s3_keys = web.ctx.site.store.get(account._key).get('s3_keys')
+    limit = RESULTS_PER_PAGE
+    offset = page * limit - limit
+    loan_history = s3_loan_api(
+        s3_keys=s3_keys,
+        action='user_borrow_history',
+        limit=limit + 1,
+        offset=offset,
+        newest=True,
+    ).json()['history']['items']
+
+    # We request limit+1 to see if there is another page of history to display,
+    # and then pop the +1 off if it's present.
+    show_next = len(loan_history) == limit + 1
+    if show_next:
+        loan_history.pop()
+
+    ocaids = [loan_record['identifier'] for loan_record in loan_history]
+    availability = get_availability_of_ocaids(ocaids)
+    loan_history_map = {
+        loan_record['identifier']: loan_record for loan_record in loan_history
+    }
+    editions = web.ctx.site.get_many(
+        [
+            '/books/%s' % availability[ocaid].get('openlibrary_edition')
+            for ocaid in availability
+            if availability[ocaid].get('openlibrary_edition')
+        ]
+    )
+
+    # Create 'placeholder' editions for items in the Internet Archive loan
+    # history, but absent from Open Library.
+    ia_only_loans = [
+        loan for loan in availability.values() if not loan.get('openlibrary_edition')
+    ]
+
+    # Attach availability and loan to both editions and ia-only items.
+    for ed in editions:
+        if ed.ocaid in ocaids:
+            ed.availability = availability.get(ed.ocaid)
+            ed.loan = loan_history_map[ed.ocaid]
+            ed.last_loan_date = ed.loan.get('updatedate')
+
+    for ia_only in ia_only_loans:
+        loan = loan_history_map[ia_only['identifier']]
+        ia_only['last_loan_date'] = loan.get('updatedate', '')
+        ia_only['ia_only'] = True  # Determines which macro to load.
+
+    editions_and_ia_loans = editions + ia_only_loans
+    editions_and_ia_loans.sort(
+        key=lambda item: item.get('last_loan_date', ''), reverse=True
+    )
+
+    result = {
+        'docs': editions_and_ia_loans,
+        'show_next': show_next,
+        'limit': limit,
+        'page': page,
+    }
+
+    return result
